[Add] Image id to open the specific image & improved Gif UI

+++ library-components/src/main/java/com/rumosoft/library_components/components/TweetContent.kt
@@ -38,7 +38,7 @@ fun TweetContent(
     images: List<ImageUI>,
     modifier: Modifier = Modifier,
     onHighlightedTextClick: (String, String) -> Unit = { _, _ -> },
-    onPictureSelected: () -> Unit = {},
+    onPictureSelected: (Long) -> Unit = {},
 ) {
     Column {
         TweetContentText(message, modifier, onHighlightedTextClick)
